Attempted to solve random test fails by restructuring the HTTP response
